requirements.txt
ipykernel
langgraph
langchain
python-dotenv
langchain-groq
langchain_tavily
langchain-core
langchain-community
langgraph-supervisor
streamlit
markdown